+++ src/main/java/com/sngular/skilltree/application/updater/ClientUpdater.java
@@ -4,7 +4,7 @@ import com.sngular.skilltree.model.Client;
 
 public interface ClientUpdater {
 
-    Client update(final Integer clientcode, final Client newClient);
+    Client update(final Long clientcode, final Client newClient);
 
-    Client patch(final Integer clientcode, final Client patchedClient);
+    Client patch(final Long clientcode, final Client patchedClient);
 }
