+++ src/main/java/com/sngular/skilltree/application/updater/PeopleUpdater.java
@@ -4,7 +4,7 @@ import com.sngular.skilltree.model.People;
 
 public interface PeopleUpdater {
 
-    People update(final Integer personcode, final People newPeople);
+    People update(final Long personcode, final People newPeople);
 
-    People patch(final Integer personcode, final People patchedPeople);
+    People patch(final Long personcode, final People patchedPeople);
 }
