+++ src/main/java/com/sngular/skilltree/contract/ProjectController.java
@@ -23,20 +23,20 @@ public class ProjectController implements ProjectApi {
     private final ProjectMapper projectMapper;
 
     @Override
-    public ResponseEntity<ProjectDTO> getProjectByCode(Integer projectcode) {
+    public ResponseEntity<ProjectDTO> getProjectByCode(Long projectcode) {
         return ResponseEntity.ok(projectMapper
                 .toProjectDTO(projectService
                         .findByCode(projectcode)));
     }
 
     @Override
-    public ResponseEntity<Void> deleteProject(Integer projectcode) {
+    public ResponseEntity<Void> deleteProject(Long projectcode) {
         var result = projectService.deleteByCode(projectcode);
         return ResponseEntity.status(result? HttpStatus.OK: HttpStatus.INTERNAL_SERVER_ERROR).build();
     }
 
     @Override
-    public ResponseEntity<ProjectDTO> updateProject(Integer projectcode, ProjectDTO projectDTO) {
+    public ResponseEntity<ProjectDTO> updateProject(Long projectcode, ProjectDTO projectDTO) {
         return ResponseEntity.ok(projectMapper
                 .toProjectDTO(projectUpdater
                         .update(projectcode, projectMapper
@@ -44,7 +44,7 @@ public class ProjectController implements ProjectApi {
     }
 
     @Override
-    public ResponseEntity<ProjectDTO> patchProject(Integer projectcode, PatchedProjectDTO patchedProjectDTO) {
+    public ResponseEntity<ProjectDTO> patchProject(Long projectcode, PatchedProjectDTO patchedProjectDTO) {
         return ResponseEntity.ok(projectMapper
                 .toProjectDTO(projectUpdater
                         .patch(projectcode, projectMapper
