+++ src/main/java/com/sngular/skilltree/model/Client.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 import java.util.List;
 
 @Builder(toBuilder = true)
-public record Client(Integer code, String name, String industry, String country, String principalOffice, String HQ,
+public record Client(Long code, String name, String industry, String country, String principalOffice, String HQ,
                      List<Office> offices, boolean deleted) {
 
 }
