+++ src/main/java/com/sngular/skilltree/model/Participate.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 import java.util.List;
 
 @Builder(toBuilder = true)
-public record Participate(Integer code, List<Roles> roles) {
+public record Participate(Long code, List<Roles> roles) {
 }
