+++ src/main/java/com/sngular/skilltree/model/People.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 
 
 @Builder(toBuilder = true)
-public record People(Integer code, String employeeId, String name, String surname, LocalDate birthDate, EnumTitle title,
+public record People(Long code, String employeeId, String name, String surname, LocalDate birthDate, EnumTitle title,
                      List<Role> roles, List<Knows> knows, List<Skill> work_with, List<Skill> master,
                      List<Participate> participate, List<Skill> interest, List<Certificate> certificates, boolean deleted) {
