+++ src/main/java/com/sngular/skilltree/model/Project.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 
 @Builder(toBuilder = true)
-public record Project(Integer code, String tag, String name, String desc, LocalDate initDate, LocalDate endDate, String domain,
+public record Project(Long code, String tag, String name, String desc, LocalDate initDate, LocalDate endDate, String domain,
                       String duration, EnumGuards guards, List<Skill> skills, List<String> historic,
                       Client client, String area, List<ProjectRoles> roles, boolean deleted) {
 }
